+++ NEWS.md
@@ -8,6 +8,7 @@ DALEX 1.0
 * added `yhat` for `lrm` models from `rms` package
 * `theme_drwhy` has now left aligned title and subtitle.
 * `residuals_distribution` calculates now diagnostic plots based on residuals ([#143](https://github.com/ModelOriented/DALEX/issues/143)).
+* `model_performance` calculated several metrics for classification and regression models ([#146](https://github.com/ModelOriented/DALEX/issues/146)).
 
 DALEX 0.4.9
 ----------------------------------------------------------------